Dispatcher: use reverse order for deletion
This patch adds support for executing the components of the Dispatcher
object in reverse order in case the object is being deleted (e.g. it has
a finalizer and a delete has been called on it).

The motivation behind is that when uninstalling a multi component object
the components might depend on each other (e.g. first you install
prometheus-operator, then the services require the the operator). In
such a setup the uninstallation should happen in reverse order to be
correct.